refactor: Encased fan now is a tileentity TileEntityFan
